requirements.txt
snowflake-connector-python[pandas]==3.12.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
snowflake-snowpark-python==1.27.0
setuptools==69.5.1
streamlit==1.45.0
plotly